documentManagement/requirements.txt
asgiref==3.7.2
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
fpdf==1.7.2
inflection==0.5.1
lxml==4.9.3
packaging==23.1
PyJWT==2.7.0
PyPDF2==3.0.1
python-docx==0.8.11
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
